Move test to MSQFaultTest
